requirements.txt
discord.py>=2.3,<3
aiohttp>=3.9,<4
python-dotenv>=1.0,<2
aiosqlite>=0.19,<1
defusedxml>=0.7,<1
pytest>=8.0
pytest-asyncio>=0.23
